FIX return value float to int convertion

+++ clic/includes/tier1/cleGaussianBlurKernel.hpp
@@ -29,7 +29,7 @@ private:
     float x;
     float y;
     float z;
-    float Sigma2KernelSize(float);
+    int Sigma2KernelSize(float);
 };
 
 } // namespace cle
